feat: remove Popup.test.tsx and ttsProviders.test.ts; add new tts-provider.test.ts and setup-test.ts
